requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
opencv-python>=4.6.0
numpy>=1.21.0
flask>=2.3.0
pillow>=9.0.0
requests>=2.28.0
huggingface-hub>=0.14.0
timm>=1.0.24
ultralytics>=8.0.0
flask-cors
flask-socketio
google-genai
openai
